requirements.txt
# === Deep Learning Framework ===
torch>=2.0.0
torchvision>=0.15.0

# === Face Detection & Alignment ===
facenet-pytorch>=2.5.3   # MTCNN
# retinaface-pytorch       # Alternative: RetinaFace (uncomment nếu dùng)

# === Image Processing ===
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0    # Advanced augmentation

# === Scientific Computing ===
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0

# === Visualization ===
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# === Config & Utilities ===
pyyaml>=6.0
tqdm>=4.65.0
easydict>=1.10

# === Dimensionality reduction (cho visualize embeddings) ===
umap-learn>=0.5.3

# === Experiment Tracking (optional nhưng khuyến nghị) ===
# wandb>=0.15.0

# === Deployment ===
# onnx>=1.14.0
# onnxruntime>=1.15.0
# streamlit>=1.25.0        # Cho demo web app
# gradio>=3.40.0           # Alternative demo

# === Testing ===
pytest>=7.4.0
pytest-cov>=4.1.0

# === Code Quality (optional) ===
# black>=23.7.0
# flake8>=6.1.0
# isort>=5.12.0
